Replace doubles with float

+++ interface/Algebra.h
@@ -10,7 +10,7 @@
 
 namespace bsm
 {
-    extern const double pi;
+    extern const float pi;
 
     LorentzVector &operator +=(LorentzVector &, const LorentzVector &);
     LorentzVector &operator -=(LorentzVector &, const LorentzVector &);
@@ -31,17 +31,17 @@ namespace bsm
     bool operator ==(const Vector &, const Vector &);
     bool operator !=(const Vector &, const Vector &);
 
-    double operator *(const LorentzVector &, const LorentzVector &);
+    float operator *(const LorentzVector &, const LorentzVector &);
 
-    double dphi(const LorentzVector &, const LorentzVector &);
-    double dr(const LorentzVector &, const LorentzVector &);
+    float dphi(const LorentzVector &, const LorentzVector &);
+    float dr(const LorentzVector &, const LorentzVector &);
 
-    double et(const LorentzVector &);
-    double eta(const LorentzVector &);
-    double mass(const LorentzVector &);
-    double momentum(const LorentzVector &);
-    double phi(const LorentzVector &);
-    double pt(const LorentzVector &);
+    float et(const LorentzVector &);
+    float eta(const LorentzVector &);
+    float mass(const LorentzVector &);
+    float momentum(const LorentzVector &);
+    float phi(const LorentzVector &);
+    float pt(const LorentzVector &);
 
     bool operator ==(const Event_Extra &, const Event_Extra &);
 }
